fix: 기존 Schedule을 NewSchedule로 대체
