+++ src/main/java/bit/schedule/controller/ScheduleController.java
@@ -1,52 +1,42 @@
 package bit.schedule.controller;
 
 import bit.schedule.dto.ScheduleRequest;
+import bit.schedule.dto.ScheduleResponse;
 import bit.schedule.service.ScheduleService;
-import bit.schedule.domain.Schedule;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
+@RequestMapping("/api/v1/schedule")
 public class ScheduleController {
-
     private final ScheduleService scheduleService;
 
-    @GetMapping("/api/v1/schedules/{id}")
-    public ResponseEntity<Schedule> getSchedule(@PathVariable Long id) {
-        Schedule schedule = scheduleService.getSchedule(id);
-
-        return ResponseEntity.ok(schedule);
+    @GetMapping("/{scheduleId}")
+    public ScheduleResponse getSchedule(@PathVariable Long scheduleId) {
+        return scheduleService.getSchedule(scheduleId);
     }
 
-    @PostMapping("/api/v1/schedules/new")
-    public ResponseEntity<Schedule> createSchedule(@Valid @RequestBody ScheduleRequest scheduleRequest) {
-        Schedule createdSchedule = scheduleService.save(scheduleRequest);
-
-        return ResponseEntity.ok(createdSchedule);
+    @GetMapping("/user/{userId}")
+    public List<ScheduleResponse> getScheduleByUser(@PathVariable Long userId) {
+        return scheduleService.getSchedulesByUserId(userId);
     }
 
-    @PatchMapping("/api/v1/schedules/{id}")
-    public ResponseEntity<Schedule> updateSchedule(@PathVariable Long id,
-                                                   @Valid @RequestBody ScheduleRequest scheduleRequest) {
-        Schedule updatedSchedule = scheduleService.patch(id, scheduleRequest);
-
-        return ResponseEntity.ok(updatedSchedule);
+    @PostMapping("")
+    public ScheduleResponse createSchedule(@Valid @RequestBody ScheduleRequest scheduleRequest) {
+        return scheduleService.saveSchedule(scheduleRequest);
     }
 
-    @DeleteMapping("/api/v1/schedules/{id}")
-    public ResponseEntity<Schedule> deleteSchedule(@PathVariable Long id) {
-        Schedule deletedSchedule = scheduleService.delete(id);
-
-        return ResponseEntity.ok(deletedSchedule);
+    @PutMapping("/{scheduleId}")
+    public ScheduleResponse updateSchedule(@PathVariable Long scheduleId, @Valid @RequestBody ScheduleRequest scheduleRequest) {
+        return scheduleService.updateSchedule(scheduleId, scheduleRequest);
     }
 
+    @DeleteMapping("/{scheduleId}")
+    public ScheduleResponse deleteSchedule(@PathVariable Long scheduleId) {
+        return scheduleService.deleteSchedule(scheduleId);
+    }
 }
