+++ src/main/java/bit/schedule/dto/ScheduleRequest.java
@@ -2,32 +2,31 @@ package bit.schedule.dto;
 
 import bit.schedule.domain.Schedule;
 import jakarta.validation.constraints.NotNull;
-import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
 
 @Getter
-@NoArgsConstructor
 public class ScheduleRequest {
 
-    @NotNull(message = "유저 아이디는 필수입니다.")
-    private String userId;
+    @NotNull(message = "유저 아이디가 필요합니다.")
+    private Long userId;
 
-    @NotNull(message = "제목은 필수입니다.")
+    @NotNull(message = "제목이 필요합니다.")
     private String title;
 
+    @NotNull(message = "내용이 필요합니다.")
     private String content;
 
-    @NotNull(message = "시작 일시는 필수입니다.")
+    @NotNull(message = "시작 일시가 필요합니다.")
     private LocalDateTime startDateTime;
 
-    @NotNull(message = "종료 일시는 필수입니다.")
+    @NotNull(message = "종료 일시가 필요합니다.")
     private LocalDateTime endDateTime;
 
     @Builder
-    public ScheduleRequest(String userId, String title, String content, LocalDateTime startDateTime,
-                           LocalDateTime endDateTime) {
+    public ScheduleRequest(Long userId, String title, String content, LocalDateTime startDateTime, LocalDateTime endDateTime) {
         this.userId = userId;
         this.title = title;
         this.content = content;
@@ -44,5 +43,4 @@ public class ScheduleRequest {
                 .endDateTime(endDateTime)
                 .build();
     }
-
 }
