+++ src/main/java/bit/schedule/repository/ScheduleRepository.java
@@ -4,7 +4,9 @@ import bit.schedule.domain.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
+
 @Repository
 public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
-
+    List<Schedule> findByUserId(Long userId);
 }
