--- src/main/java/bit/schedule_new/service/NewScheduleService.java
@@ -1,18 +1,0 @@
-package bit.schedule_new.service;
-
-import bit.schedule_new.dto.NewScheduleRequest;
-import bit.schedule_new.dto.NewScheduleResponse;
-
-import java.util.List;
-
-public interface NewScheduleService {
-    NewScheduleResponse getSchedule(Long scheduleId);
-
-    List<NewScheduleResponse> getSchedulesByUserId(Long userId);
-
-    NewScheduleResponse saveSchedule(NewScheduleRequest newScheduleRequest);
-
-    NewScheduleResponse updateSchedule(Long scheduleId, NewScheduleRequest newScheduleRequest);
-
-    NewScheduleResponse deleteSchedule(Long scheduleId);
-}
