+++ src/test/java/bit/schedule/domain/ScheduleTest.java
@@ -1,40 +1,33 @@
 package bit.schedule.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import bit.schedule.dto.ScheduleRequest;
-import java.time.LocalDateTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import java.time.LocalDateTime;
+
+import static bit.schedule.util.ScheduleRequestFixture.getNewScheduleRequest;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
 class ScheduleTest {
 
-    @DisplayName("schedule 객체를 수정 한다.")
+    @DisplayName("스케줄 객체를 수정 한다.")
     @Test
     void updateTest() {
-        // given
-        ScheduleRequest scheduleRequest = ScheduleRequest.builder()
-                .userId("testId")
-                .title("일정 제목")
-                .content("일정 내용")
-                .startDateTime(LocalDateTime.of(2021, 10, 10, 10, 10))
-                .endDateTime(LocalDateTime.of(2021, 10, 10, 10, 10))
-                .build();
-
+        //Given
+        ScheduleRequest scheduleRequest = getNewScheduleRequest(LocalDateTime.now(), LocalDateTime.now().plusHours(1));
         Schedule schedule = scheduleRequest.toEntity();
-
-        ScheduleRequest updatedScheduleRequest = ScheduleRequest.builder()
-                .userId("testId")
-                .title("수정된 일정 제목")
-                .content("수정된 일정 내용")
-                .startDateTime(LocalDateTime.of(2022, 10, 10, 10, 10))
-                .endDateTime(LocalDateTime.of(2022, 10, 10, 10, 10))
+        ScheduleRequest updateScheduleRequest = ScheduleRequest.builder()
+                .userId(2L)
+                .title("title update test")
+                .content("content update test")
+                .startDateTime(LocalDateTime.of(2024, 9, 1, 1, 1))
+                .endDateTime(LocalDateTime.of(2024, 9, 1, 1, 1))
                 .build();
-
-        // when
-        schedule.update(updatedScheduleRequest);
-
-        // then
+        //When
+        schedule.update(updateScheduleRequest);
+        //Then
         assertThat(schedule).extracting(
                 "userId",
                 "title",
@@ -42,12 +35,32 @@ class ScheduleTest {
                 "startDateTime",
                 "endDateTime"
         ).containsExactly(
-                "testId",
-                "수정된 일정 제목",
-                "수정된 일정 내용",
-                LocalDateTime.of(2022, 10, 10, 10, 10),
-                LocalDateTime.of(2022, 10, 10, 10, 10)
+                2L,
+                "title update test",
+                "content update test",
+                LocalDateTime.of(2024, 9, 1, 1, 1),
+                LocalDateTime.of(2024, 9, 1, 1, 1)
         );
+    }
+
+    @DisplayName("스케줄 객체를 수정 시 시작 시간이 종료 시간보다 늦으면 에러를 발생시킨다.")
+    @Test
+    void updateValidStartEndTimeTest() {
+        //Given
+        ScheduleRequest scheduleRequest = getNewScheduleRequest(LocalDateTime.now(), LocalDateTime.now().plusHours(1));
+        Schedule schedule = scheduleRequest.toEntity();
+        ScheduleRequest updateScheduleRequest = ScheduleRequest.builder()
+                .userId(2L)
+                .title("title update test")
+                .content("content update test")
+                .startDateTime(LocalDateTime.of(2024, 9, 1, 1, 1))
+                .endDateTime(LocalDateTime.of(2024, 8, 1, 1, 1))
+                .build();
+        //When
+        //Then
+        assertThatThrownBy(() -> schedule.update(updateScheduleRequest))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("시작 시간은 종료 시간보다 늦을 수 없습니다.");
     }
 
 }
